// src/redux/actions/action-types.ts
export enum ActionTypes {
  FETCH_USERS = "FETCH_USERS",
  USERS_LIMIT_REACHED = "USERS_LIMIT_REACHED",
  UPDATE_USERS = "UPDATE_USERS",
  FETCH_POSTS = "FETCH_POSTS",
  POSTS_LIMIT_REACHED = "POSTS_LIMIT_REACHED",
  FETCH_TODOS = "FETCH_TODOS",
  TODOS_LIMIT_REACHED = "TODOS_LIMIT_REACHED",
  MODAL_OPEN = "MODAL_OPEN",
  MODAL_CLOSE = "MODAL_CLOSE",
}